Separated sending and receiving logic from core Session. Added new SendHeaderWithBody to allow for sending direct byte sequences.
