+++ src/NexNet/Internals/NexNetSession.Sending.cs
@@ -1,0 +1,141 @@
+﻿using System.Buffers;
+using NexNet.Messages;
+using NexNet.Invocation;
+using System.IO.Pipelines;
+using System.Threading.Tasks;
+using System.Threading;
+using System;
+using MemoryPack;
+
+namespace NexNet.Internals;
+
+internal partial class NexNetSession<THub, TProxy> : INexNetSession<TProxy>
+    where THub : HubBase<TProxy>, IMethodInvoker<TProxy>, IInvocationMethodHash
+    where TProxy : ProxyInvocationBase, IProxyInvoker, IInvocationMethodHash, new()
+{
+   
+    public async ValueTask SendHeaderWithBody<TMessage>(TMessage body, CancellationToken cancellationToken = default)
+        where TMessage : IMessageBodyBase
+    {
+        if (_pipeOutput == null || cancellationToken.IsCancellationRequested)
+            return;
+
+        if (State != ConnectionState.Connected && State != ConnectionState.Disconnecting)
+            return;
+
+        using var mutexResult = await _writeMutex.TryWaitAsync(cancellationToken).ConfigureAwait(false);
+
+        var header = _bufferWriter.GetMemory(3);
+        _bufferWriter.Advance(3);
+        MemoryPackSerializer.Serialize(_bufferWriter, body);
+
+        var contentLength = checked((ushort)(_bufferWriter.Length - 3));
+
+        header.Span[0] = (byte)TMessage.Type;
+
+        BitConverter.TryWriteBytes(header.Span.Slice(1, 2), contentLength);
+
+        var length = (int)_bufferWriter.Length;
+
+        using var buffer = _bufferWriter.Flush();
+
+        _config.InternalOnSend?.Invoke(this, buffer.Value.ToArray());
+
+        buffer.Value.CopyTo(_pipeOutput.GetSpan(length));
+        _pipeOutput.Advance(length);
+
+        _config.Logger?.LogTrace($"Sending {TMessage.Type} message & body with {length} bytes.");
+
+        var result = await _pipeOutput.FlushAsync(cancellationToken).ConfigureAwait(false);
+
+        OnSent?.Invoke();
+
+        if (result.IsCanceled || result.IsCompleted)
+            await DisconnectCore(DisconnectReason.ProtocolError, false).ConfigureAwait(false);
+    }
+
+
+    public async ValueTask SendHeaderWithBody(MessageType type, ReadOnlySequence<byte> body, CancellationToken cancellationToken = default)
+    {
+        if (_pipeOutput == null || cancellationToken.IsCancellationRequested)
+            return;
+
+        if (State != ConnectionState.Connected && State != ConnectionState.Disconnecting)
+            return;
+
+        using var mutexResult = await _writeMutex.TryWaitAsync(cancellationToken).ConfigureAwait(false);
+
+        if (mutexResult.Success != true)
+            throw new InvalidOperationException("Could not acquire write lock");
+        var length = (int)body.Length;
+        var contentLength = checked((ushort)(length));
+
+        var header = _pipeOutput.GetMemory(3);
+        header.Span[0] = (byte)type;
+        BitConverter.TryWriteBytes(header.Span.Slice(1, 2), contentLength);
+        _pipeOutput.Advance(3);
+        body.CopyTo(_pipeOutput.GetSpan((int)body.Length));
+        _pipeOutput.Advance(length);
+
+        if (_config.InternalOnSend != null)
+        {
+            var debugCopy = new byte[body.Length + 3];
+            debugCopy[0] = (byte)type;
+            BitConverter.TryWriteBytes(new Span<byte>(debugCopy).Slice(1, 2), contentLength);
+            body.CopyTo(new Span<byte>(debugCopy).Slice(3));
+            _config.InternalOnSend?.Invoke(this, debugCopy);
+        }
+
+        _config.Logger?.LogTrace($"Sending {length} bytes with header and length.");
+        FlushResult result = default;
+        try
+        {
+            result = await _pipeOutput.FlushAsync(cancellationToken).ConfigureAwait(false);
+        }
+        catch (ObjectDisposedException)
+        {
+
+        }
+
+        OnSent?.Invoke();
+
+        if (result.IsCanceled || result.IsCompleted)
+            await DisconnectCore(DisconnectReason.ProtocolError, false).ConfigureAwait(false);
+    }
+
+
+    public async ValueTask SendHeader(MessageType type, CancellationToken cancellationToken = default)
+    {
+        if (_pipeOutput == null || cancellationToken.IsCancellationRequested)
+            return;
+
+        if (State != ConnectionState.Connected && State != ConnectionState.Disconnecting)
+            return;
+
+        using var mutexResult = await _writeMutex.TryWaitAsync(cancellationToken).ConfigureAwait(false);
+
+        if (mutexResult.Success != true)
+            throw new InvalidOperationException("Could not acquire write lock");
+
+        _config.InternalOnSend?.Invoke(this, new[] { (byte)type });
+
+        _pipeOutput.GetSpan(1)[0] = (byte)type;
+        _pipeOutput.Advance(1);
+
+        _config.Logger?.LogTrace($"Sending {type} header.");
+        FlushResult result = default;
+        try
+        {
+            result = await _pipeOutput.FlushAsync(cancellationToken).ConfigureAwait(false);
+        }
+        catch (ObjectDisposedException)
+        {
+
+        }
+
+        OnSent?.Invoke();
+
+        if (result.IsCanceled || result.IsCompleted)
+            await DisconnectCore(DisconnectReason.ProtocolError, false).ConfigureAwait(false);
+    }
+}
